organized the files a little bit
