+++ src/components/App/App.css
@@ -24,9 +24,30 @@
   font-size: calc(10px + 2vmin);
   color: white;
   overflow-y: scroll;
-
 }
 
 .App-link {
   color: #61dafb;
+}
+
+#chat-form {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-evenly;
+  align-items: baseline;
+  position: fixed;
+  bottom: 0;
+  left: 10px;
+}
+
+input[type="text"] {
+  color: #ffffff;
+}
+
+#chat-form label {
+  color: #ffffff;
+}
+
+#standard-basic {
+  width: 270px;
 }
