+++ src/components/MessageForm/MessageForm.js
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import React from "react";
 import { TextField, IconButton } from "@material-ui/core";
 import "./MessageForm.css";
 import SendIcon from "@material-ui/icons/Send";
