+++ src/components/index.js
@@ -1,6 +1,6 @@
-import App from "./App/App";
-import SimpleMenu from "./Menu/Menu";
-import MessageForm from "./MessageForm/MessageForm";
-import Messages from "./Messages/Messages";
-import Navbar from "./Navbar/Navbar";
-import { default as Signup } from "./SignupForm/SignupForm";
+export { default as App } from "./App/App";
+export { default as SimpleMenu } from "./Menu/Menu";
+export { default as MessageForm } from "./MessageForm/MessageForm";
+export { default as Messages } from "./Messages/Messages";
+export { default as Navbar } from "./Navbar/Navbar";
+export { default as Signup } from "./SignupForm/SignupForm";
